+++ examples/example01/__init__.py
@@ -1,0 +1,6 @@
+"""
+Author: L. Saetta
+Date last modified: 2026-07-21
+License: MIT
+Description: Basic LangGraph flow persisted with an Oracle ADB checkpointer.
+"""